refactor(workflow/engine): modularize engine.rs into submodules
Split the 1397-line engine/mod.rs into 6 focused files by domain concern:

- engine/mod.rs: struct definition, start/with_event_broadcaster/store/broadcast,
  helpers (strip_continued_suffix, timestamp_now, inject_engine_version,
  ENGINE_VERSION), and engine_version_stamp_tests. Submodule declarations added.
  store and event_tx fields marked pub(crate); helpers marked pub(crate) so
  sibling impl blocks can call them.

- engine/workflows.rs: workflow lifecycle — start_workflow, get_workflow,
  list_workflows, upsert_search_attributes, cancel_workflow, terminate_workflow,
  finalise_cancellation, complete_workflow, fail_workflow,
  notify_parent_of_child_outcome (private).

- engine/children.rs: child workflows and continue-as-new — start_child_workflow,
  list_child_workflows, continue_as_new.

- engine/activities.rs: activity ops and side effects — schedule_activity,
  claim_activity, get_activity, complete_activity, fail_activity,
  heartbeat_activity, record_side_effect.

- engine/tasks.rs: workflow-task dispatch and timers — claim_workflow_task,
  submit_workflow_commands, schedule_timer.

- engine/collections.rs: grouped small ops — send_signal, get_events,
  register_worker, heartbeat_worker, list_workers, create_schedule,
  list_schedules, get_schedule, delete_schedule, update_schedule,
  set_schedule_paused, create_namespace, list_namespaces, delete_namespace,
  get_namespace_stats, get_queue_stats, create_snapshot, get_latest_snapshot.

No method bodies or signatures changed. Tests: 1092 passed / 0 failed / 1 ignored.
